abstract away parsing messages from pi
