+++ pi/src/communication/initialization.rs
@@ -1,27 +1,15 @@
 use anyhow::{Context as AnyhowContext, Error};
 use async_nats::{
-    Client, ConnectOptions, Message,
+    Client, ConnectOptions,
     jetstream::{
         self,
         stream::{self, Info},
     },
 };
-use futures::{Stream, future::try_join_all, stream::select_all};
-use once_cell::sync::Lazy;
 use std::env;
 
 use super::Communication;
 use crate::CONFIG;
-
-const PLUG_TOPICS: &[&str] = &["stat.*.RESULT", "stat.*.STATUS8"];
-const SOLAREDGE_TOPICS: &[&str] = &["solaredge.modbus.battery.battery0", "solaredge.powerflow"];
-
-static RELEVANT_SMART_HOME_TOPICS: Lazy<Vec<&str>> = Lazy::new(|| {
-    let mut topics = Vec::new();
-    topics.extend_from_slice(PLUG_TOPICS);
-    topics.extend_from_slice(SOLAREDGE_TOPICS);
-    topics
-});
 
 impl Communication {
     pub async fn connect() -> Result<Self, Error> {
@@ -42,26 +30,6 @@ impl Communication {
             .await
             .context("Could not create the state stream for the service")
     }
-
-    pub async fn subscribe_to_smart_home(&self) -> Result<impl Stream<Item = Message>, Error> {
-        nats_subscribe(self.pi_nats.clone(), RELEVANT_SMART_HOME_TOPICS.to_vec())
-            .await
-            .context("Could not subscribe to the subjects on the controller")
-    }
-}
-
-async fn nats_subscribe(
-    nats: Client,
-    subjects: Vec<&str>,
-) -> Result<impl Stream<Item = Message>, Error> {
-    let subscribers = try_join_all(
-        subjects
-            .iter()
-            .map(async |subject| nats.subscribe(subject.to_string()).await),
-    )
-    .await?;
-
-    Ok(select_all(subscribers))
 }
 
 async fn connect_nats_client(prefix: &str) -> Result<Client, Error> {
