+++ pi/src/communication/mod.rs
@@ -2,9 +2,22 @@ pub mod events;
 mod initialization;
 
 use anyhow::{Context as AnyhowContext, Error};
-use async_nats::{Client, jetstream::Context};
+use async_nats::{Client, Message, jetstream::Context};
+use events::UpdateEvent;
+use futures::{Stream, StreamExt, future::try_join_all, stream::select_all};
+use once_cell::sync::Lazy;
 
 use crate::CONFIG;
+
+const PLUG_TOPICS: &[&str] = &["stat.*.RESULT", "stat.*.STATUS8"];
+const SOLAREDGE_TOPICS: &[&str] = &["solaredge.modbus.battery.battery0", "solaredge.powerflow"];
+
+static RELEVANT_SMART_HOME_TOPICS: Lazy<Vec<&str>> = Lazy::new(|| {
+    let mut topics = Vec::new();
+    topics.extend_from_slice(PLUG_TOPICS);
+    topics.extend_from_slice(SOLAREDGE_TOPICS);
+    topics
+});
 
 #[derive(Debug, Clone)]
 pub struct Communication {
@@ -13,6 +26,19 @@ pub struct Communication {
 }
 
 impl Communication {
+    pub async fn get_update_events(
+        &self,
+    ) -> Result<impl Stream<Item = Result<UpdateEvent, Error>>, Error> {
+        let pi_messages = self.subscribe_to_pi().await?;
+        Ok(pi_messages.map(|message| UpdateEvent::try_from(&message)))
+    }
+
+    async fn subscribe_to_pi(&self) -> Result<impl Stream<Item = Message>, Error> {
+        nats_subscribe(self.pi_nats.clone(), RELEVANT_SMART_HOME_TOPICS.to_vec())
+            .await
+            .context("Could not subscribe to the subjects on the controller")
+    }
+
     pub async fn query_plug_state(&self) -> Result<(), Error> {
         self.pi_nats
             .publish(
@@ -45,4 +71,18 @@ impl Communication {
 
         Ok(())
     }
+}
+
+async fn nats_subscribe(
+    nats: Client,
+    subjects: Vec<&str>,
+) -> Result<impl Stream<Item = Message>, Error> {
+    let subscribers = try_join_all(
+        subjects
+            .iter()
+            .map(async |subject| nats.subscribe(subject.to_string()).await),
+    )
+    .await?;
+
+    Ok(select_all(subscribers))
 }
